Fix an issue with non-built-in default devices
When no audio devices for input, output, or ringtone have been
selected, we want to use first met built-in devices with appropriate
capabilities (with non-zero number of input channels for import, and
similar rule for output).

Originally, the OS API didn't allow to tell if the device is
built-in. Not it allows.

Fixes #71