Use separate methods for getting by name for input and output
Some devices are listed in the system twice. One time for input only,
another time for output only. When only one name to device mapping is
used, it can lead to an error. The map would then return the same
audio device entry in the system for both, input only and output only
device, because they have the same name.

Issue #145