Fix object count calculation
When using CoreAudio API, the length of the expected data is received
separately from receiving the data itself. Furthermore, the expected
type is not part of the function signature, but is just specified in the
documentation.

Our code expected that the data length it receives always divides
to the size of the expected data type without remainder. But when
virtual audio devices are configured in the system (for example, using
the Audio MIDI Setup app) that is not the case.

Issue #221