Create ABAddressBook before each enumeration
According to the documentation, the (not shared) ABAddressBook
instance is usually used for short-time purposes, and the changes to
the user's address book are not automatically merged into it. So in
our case, it should be created before each enumeration and discarded
to contain the recent changes each time.

Issue #238