Get call info within callback before dispatching to main
A race condition occurred when only the call id was passed to main
thread. Sometimes, the call to pjsua_call_get_info() was made too
late, when the PJSIP call didn't exist any more. Now, the call info is
got within the callback in backgroun and the whole call info structure
is passed to the main thread.

Issue #433