Create autorelease pools around background execution
The documentation on threads and runloops states that an autorelease
pool is created automatically only on the main thread. It also happens
in practice at least on macOS 10.10 and later. But because the
documentation makes no guarantees, explicit autorelease pools should
be created around calls on a long-living background thread.